Create ZipSHP
Script zips all individual shapefiles within a defined directory. Zipfiles are named identical to the shapefile itself. This is useful for platforms like CartoDB and MapBox, especially if you are in an Esri workflow. If you are usging a geodatabase . If the destination doesn’t exist, it will be created by the script. See:  http://www.bgcarto.com/zip-all-shapefiles-in-directory-individually/